--- project/src/App.js
@@ -1,7 +1,0 @@
-function App() {
-  return (
-    null
-  );
-}
-
-export default App;
