+++ project/src/Routes.js
@@ -1,0 +1,15 @@
+import { BrowserRouter as Router, Route, Routes} from "react-router-dom"
+import MainPage from "./routes/MainPage";
+import NavBar from "./routes/Navbar";
+
+function Routers() {
+    return(
+        <Router>
+            <NavBar />
+            <Routes>
+                <Route path="/" element={<MainPage />} />
+            </Routes>
+        </Router>
+    )
+}
+export default Routers;
