+++ project/src/routes/MainPage.jsx
@@ -1,0 +1,4 @@
+function MainPage () {
+    return(<div>메인페이지</div>)
+}
+export default MainPage;
