+++ project/src/routes/Navbar.jsx
@@ -1,0 +1,5 @@
+function NavBar () {
+    return(<div>navbar</div>)
+}
+
+export default NavBar;
